legacy_backup/requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
python-dateutil==2.8.2
pytz==2023.3
groq==0.4.1
gunicorn==21.2.0
